few updates for gan easy run
