+++ catalyst/core/callback.py
@@ -1,6 +1,7 @@
 from enum import IntFlag
 from typing import Callable, List
 
+from catalyst import utils
 from .state import _State
 
 
@@ -85,6 +86,20 @@ class LoggerCallback(MasterOnlyCallback):
         if order is None:
             order = CallbackOrder.Internal
         super().__init__(order=order)
+
+
+class RaiseExceptionCallback(LoggerCallback):
+    def __init__(self):
+        order = CallbackOrder.Other + 1
+        super().__init__(order=order)
+
+    def on_exception(self, state: _State):
+        exception = state.exception
+        if not utils.is_exception(exception):
+            return
+
+        if state.need_reraise_exception:
+            raise exception
 
 
 class MetricCallback(Callback):
